Move XDG get_user_*_dir() functions from microdnf to libdnf utils
The functions return directory path according to the XDG Base Directory
Specification.
