+++ microdnf/utils.hpp
@@ -23,7 +23,6 @@ along with microdnf.  If not, see <https://www.gnu.org/licenses/>.
 #include <libdnf/base/goal.hpp>
 #include <sys/types.h>
 
-#include <filesystem>
 #include <string>
 
 namespace microdnf {
@@ -35,30 +34,6 @@ bool am_i_root() noexcept;
 /// The getuid() is returned instead if there was a problem.
 /// The value is cached.
 uid_t get_login_uid() noexcept;
-
-namespace xdg {
-
-/// Returns user home directory.
-std::filesystem::path get_user_home_dir();
-
-/// Returns user cache directory.
-/// A base directory relative to which user specific non-essential data files should be stored.
-std::filesystem::path get_user_cache_dir();
-
-/// Returns user configuration directory.
-/// A base directory relative to which user specific configuration files should be stored.
-std::filesystem::path get_user_config_dir();
-
-/// Returns user data directory.
-/// A base directory relative to which user specific data files should be stored.
-std::filesystem::path get_user_data_dir();
-
-/// Returns user runtime directory.
-/// A base directory relative to which user-specific non-essential runtime files and other file objects
-/// (such as sockets, named pipes, ...) should be stored.
-std::filesystem::path get_user_runtime_dir();
-
-}  // namespace xdg
 
 /// find the base architecture
 const char * get_base_arch(const char * arch);
